Revert "Impl 'Delta Computation' (#8)"
This reverts commit 0941371250c4d10ba23a30444d74fd12fb90ae75.
