--- data/saved_data/auth_export/accounts.json
@@ -1,1 +1,0 @@
-{"kind":"identitytoolkit#DownloadAccountResponse","users":[{"localId":"JOKYpMbqVxruIKO5pj9jvz4pySAr","createdAt":"1616120233889","lastLoginAt":"1616670602347","displayName":"John McClane","photoUrl":"https://pbs.twimg.com/profile_images/2647803927/1f3436521ba2ee011d4e6a9d60596c44.jpeg","providerUserInfo":[{"providerId":"google.com","rawId":"0861056242461056649423093128039461622495","federatedId":"0861056242461056649423093128039461622495","displayName":"John McClane","photoUrl":"https://pbs.twimg.com/profile_images/2647803927/1f3436521ba2ee011d4e6a9d60596c44.jpeg","email":"john@example.com","screenName":"John"}],"validSince":"1616502129","email":"john@example.com","emailVerified":true,"disabled":false,"lastRefreshAt":"2021-03-25T11:10:02.347Z"},{"localId":"OJ5WOiv3hU07LQzNMyGj8jB0Kgxc","createdAt":"1616269116330","lastLoginAt":"1616670607558","displayName":"Diana Prince","photoUrl":"https://news.artnet.com/app/news-upload/2019/12/Movies-Gallery_WW84_05_5deb0dd8d64557.66209613-256x256.jpg","providerUserInfo":[{"providerId":"google.com","rawId":"5241301171256853587898344326373797001945","federatedId":"5241301171256853587898344326373797001945","displayName":"Diana Prince","photoUrl":"https://news.artnet.com/app/news-upload/2019/12/Movies-Gallery_WW84_05_5deb0dd8d64557.66209613-256x256.jpg","email":"diana@example.com","screenName":"Wonder Woman"}],"validSince":"1616502129","email":"diana@example.com","emailVerified":true,"disabled":false,"lastRefreshAt":"2021-03-25T18:02:27.270Z"}]}
