--- data/saved_data/auth_export/config.json
@@ -1,1 +1,0 @@
-{"signIn":{"allowDuplicateEmails":false}}
