--- data/saved_data/firebase-export-metadata.json
@@ -1,12 +1,0 @@
-{
-  "version": "9.6.1",
-  "firestore": {
-    "version": "1.11.12",
-    "path": "firestore_export",
-    "metadata_file": "firestore_export/firestore_export.overall_export_metadata"
-  },
-  "auth": {
-    "version": "9.6.1",
-    "path": "auth_export"
-  }
-}
