+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "captains-log",
+  "name": "lexer",
   "version": "0.0.1",
-  "description": "Captain's Log helps keep track of your days, months, and years!",
+  "description": "My lexer writing program.",
   "private": true,
   "dependencies": {
     "firebase": "^8.2.9",
@@ -35,23 +35,23 @@
     "typescript": "^4.1.3",
     "webpack": "^5.24.2",
     "webpack-cli": "^4.5.0",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-merge": "^5.7.3"
+    "webpack-dev-server": "^3.11.2"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --config webpack.prod.js",
-    "start": "webpack serve --config webpack.dev.js",
+    "build": "webpack",
+    "watch": "webpack --watch",
+    "start": "webpack serve",
     "format": "prettier --write \"src/**/*.{js,jsx}\""
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/r002/captains-log.git"
+    "url": "git+https://github.com/r002/lexer.git"
   },
   "author": "Robert Lin",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/r002/captains-log/issues"
+    "url": "https://github.com/r002/lexer/issues"
   },
-  "homepage": "https://captains-log.web.app"
+  "homepage": "https://github.com/r002/lexer#readme"
 }
