+++ src/index.html
@@ -2,26 +2,13 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <title>#8: Setup webpack.prod.js</title>
+    <title>#8: Delta Computation</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
   </head>
   <body style="visibility: hidden;">
     <div id="root"></div>
     <br /><br /><br /><br /><br /><br /><br />
     <hr />
-    <h2>Project Motivation:</h2>
-    Several motivations inspired and drove this project.
-    <br /><br />
-    First: I've always been a big fan of the "self-quantification movement." I own a Fitbit and enjoy seeing granular metrics on everything I do: Sleep, exercise, etc. But Fitbit alone was insufficient. I wanted to know how much time I spent eating, brushing my teeth, running errands, etc.
-    <br /><br />
-    Second: I wanted a good project with which to learn React with hooks and what webdev looks like nowadays in 2021. Webpack, ESLint, TypeScript, Firebase preview channels, the whole nine yards. My previous knowledge on the webdev front was horribly outdated. Also: Tests, design patterns, Git. It was high time to get caught up.
-    <br /><br />
-    Finally, realistically though, what ***really*** drove this project: My wife, honest to God, bugs me probably once a week that "I don't spend enough time with her." And that: "I love my computer and spend more time with it than I love her."  Utter baloney.  With this app, I now finally have hard, cold data to show The Wife that, "See? We *did* visit the park four times last week!" and "I spent an aggregate total of __ hours __ minutes with you last month watching Netflix, going out for walks, and eating."
-    <br /><br />
-    Data and technology to the rescue! 😀
-    <br /><br />
-    <hr />
-    <br />
     <h2>Release History:</h2>
     <ul>
       <li><a href='https://github.com/r002/captains-log/pull/8'>Tue - 3/23/21 | v0.0.8 | Add Delta Computation</a></li>
