+++ src/style.css
@@ -7,7 +7,6 @@
 
 body {
   background-color: #81a69b;
-  /* background-color: #151515; */
   margin: 0;
   padding: 0;
 }
