+++ src/widgets/LogEntry.tsx
@@ -1,9 +1,7 @@
-import styled from 'styled-components'
+import styled, { css } from 'styled-components'
 import React, { useEffect, useState, useContext } from 'react'
 import { UserContext } from '../providers/AuthContext'
 import firebase from 'firebase/app'
-import { FormattedDt, ILog } from './Shared'
-import { LogViewer } from './LogViewer'
 
 const Box = styled.div`
   padding: 20px;
@@ -21,6 +19,14 @@ const Prompt = styled.div`
   margin-bottom: 10px;
 `
 
+const LimeGreen = styled.span`
+  color: limegreen;
+`
+
+const Yellow = styled.span`
+  color: yellow;
+`
+
 const LogInput = styled.input`
   width: 900px;
   font-size: 22px;
@@ -29,6 +35,44 @@ const LogInput = styled.input`
   border: none;
   outline: none;
 `
+// Awaiting TC39 to approve 'Temporal'! 🥴
+const formattedDt = (d: Date) => {
+  return (
+    <><LimeGreen>{d.toString().slice(0, 15)} </LimeGreen>
+      <Yellow>{d.toString().slice(16, 24)} ET</Yellow></>
+      // <Yellow>{d.toString().slice(16, 21)} ET</Yellow> // Omit seconds
+  )
+}
+
+interface IFlog {
+  readonly background?: string
+}
+
+const FLogRecord = styled.div<IFlog>`
+  padding: 10px;
+  width: 900px;
+  background: darkblue;
+  color: white;
+  box-sizing: border-box;
+  border: solid darkgray 1px;
+
+  ${props => props.background && css`
+    background: ${props.background};
+  `}
+`
+interface ILog {
+  dt: Date
+  activity: string
+}
+
+const LogRecord = ({ dt, activity, saved }: {dt: Date, activity: string, saved: boolean}) => {
+  const bg = saved ? 'darkblue' : '#00468b'
+  return (
+    <FLogRecord title={dt.toString()} background={bg}>
+      {formattedDt(dt)} :: {activity}
+    </FLogRecord>
+  )
+}
 
 async function getLogs (user: firebase.User) : Promise<Array<ILog>> {
   console.log('----------------- fire getLogs!')
@@ -58,6 +102,7 @@ export const LogEntry = () => {
   // console.log('🚀🚀 LogEntry BEGIN rendering')
 
   const { user } = useContext(UserContext)
+  const [logsFromDb, setLogsFromDb] = useState([] as Array<ILog>)
   const [logs, setLogs] = useState([] as Array<ILog>)
   const [dt, setDt] = useState(new Date())
   const [activity, setActivity] = useState('')
@@ -77,28 +122,22 @@ export const LogEntry = () => {
   }
 
   // Initial load logs from db if user signs in
-  useEffect(() => {
-    // console.log('******** LogEntry fire useEffect', user)
-    if (user) {
-      getLogs(user).then(logsFromDb => {
-        if (logs.length === 0 || logs.length + logsFromDb.length > logsFromDb.length) {
-          // If user has written logs anonymously, first write the unsaved logs to Firestore
-          for (const log of logs) {
-            writeLog(user, log)
-            console.log('>> write unsaved log to firestore', log)
-          }
-          // Now update the logs with any from the db. This will update the view.
-          setLogs(oldLogs => [
-            ...oldLogs,
-            ...logsFromDb
-          ])
-        }
-      })
-    } else {
-      // console.log('&&&&& LogEntry User has logged out; clear logs', user)
-      setLogs([])
-    }
-  }, [user])
+  if (user && logsFromDb.length === 0) {
+    getLogs(user).then(l => {
+      setLogsFromDb(l)
+
+      // If user has written logs anonymously, write the unsaved logs to Firestore
+      for (const log of logs) {
+        writeLog(user, log)
+        console.log('>> write unsaved log to firestore', log)
+      }
+    })
+
+  // Clear logs from db if user signs out
+  } else if (!user && logsFromDb.length > 0) {
+    setLogsFromDb([])
+    setLogs([])
+  }
 
   function addLog (e: React.KeyboardEvent<HTMLInputElement>) {
     if (e.key === 'Enter') {
@@ -123,13 +162,15 @@ export const LogEntry = () => {
   return (
     <>
       <Box>
-        <Prompt><span title='Robert Shell 😄'>RS</span> {FormattedDt(dt)}</Prompt>
+        <Prompt><span title='Robert Shell 😄'>RS</span> {formattedDt(dt)}</Prompt>
         $ <LogInput autoFocus={true} value={activity} onChange={handleChange} onKeyDown={addLog}></LogInput>
       </Box>
       <br /><br />
       <hr />
       <br /><br />
-      <LogViewer logs={logs} />
+      {/* {logs} */}
+      {logs.map((l: ILog, i: number) => <LogRecord key={'log' + i} {...l} saved={!!user} />)}
+      {logsFromDb.map((l: ILog, i: number) => <LogRecord key={'log' + i} {...l} saved={true} />)}
     </>
   )
 }
