+++ webpack.config.js
@@ -1,17 +1,32 @@
 const path = require('path')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
+// const { CleanWebpackPlugin } = require('clean-webpack-plugin')
 
 module.exports = {
+  mode: 'development',
   entry: {
     app: './src/index.tsx'
   },
+  devtool: 'inline-source-map',
+  devServer: {
+    contentBase: './dist',
+    hot: true,
+    // inline: false,
+    port: 3000
+  },
   plugins: [
     new HtmlWebpackPlugin({
-      title: 'Production',
       template: './src/index.html',
       favicon: './src/favicon.ico'
     })
   ],
+  // plugins: [
+  //   // new CleanWebpackPlugin(['dist/*']) for < v2 versions of CleanWebpackPlugin
+  //   new CleanWebpackPlugin(),
+  //   new HtmlWebpackPlugin({
+  //     title: 'Hot Module Replacement'
+  //   })
+  // ],
   module: {
     rules: [
       {
@@ -31,8 +46,7 @@ module.exports = {
     extensions: ['.tsx', '.ts', '.js']
   },
   output: {
-    filename: '[name].bundle.js',
-    path: path.resolve(__dirname, 'dist'),
-    clean: true
+    filename: 'bundle.js',
+    path: path.resolve(__dirname, 'dist')
   }
 }
