--- webpack.prod.js
@@ -1,6 +1,0 @@
-const { merge } = require('webpack-merge')
-const common = require('./webpack.common.js')
-
-module.exports = merge(common, {
-  mode: 'production'
-})
